Move encoding integration tests to tests/encodings.rs
